backend/requirements.txt
# Python 3.10+ Recommended
# --- CENTRAL INTELLIGENCE SERVER ---
fastapi==0.111.0
uvicorn[standard]==0.29.0
python-multipart==0.0.9
sse-starlette==1.8.2
requests==2.31.0
aiohttp==3.9.5

# --- AUTH & SECURITY ---
bcrypt==4.1.3
python-jose[cryptography]==3.3.0

# --- AI & VISION ---
opencv-python==4.9.0.80
insightface==0.7.3
ultralytics==8.2.0
onnxruntime==1.18.0
onnxruntime-gpu==1.18.0 
numpy==1.26.4
Pillow==10.3.0

# --- DATABASE & STORAGE ---
qdrant-client==1.9.1
aiohttp
static-ffmpeg==2.5
